// src/modules/hr/types/position.ts
export interface PositionExpListDto extends BaseDto {
  positionId: UUID;
  samePosExp: number;
  otherPosExp: number;
  minAge: number;
  maxAge: number;
}

export interface PositionExpAddDto {
  samePosExp: number;
  otherPosExp: number;
  minAge: number;
  maxAge: number;
  positionId: UUID;
}

export interface PositionExpModDto {
  id: UUID;
  samePosExp: number;
  otherPosExp: number;
  minAge: number;
  maxAge: number;
  positionId: UUID;
  rowVersion: string;
}
// ...
export interface BaseDto {
  id: UUID;
  isDeleted: boolean;
  rowVersion: string;
  dateAdd: string;
  dateMod: string | null;
}